Revert "Merge branch 'main' of https://github.com/IlliaMyronov/MissileCommand"
This reverts commit 7d0dbc99871626d6679fdf1010585680a6039af3, reversing
changes made to f9ffcf703e8b402ff1da0eb75b516e714ebf14f4.

+++ game/God Simulator/Assets/Scripts/TileScripts/TileInfo.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class TileInfo : MonoBehaviour
 {
     [SerializeField] private int growthDifficulty;
-    [SerializeField] private int moveDifficulty;
+    [SerializeField] private float moveDifficulty;
     [SerializeField] private List<GameObject> growableObjects;
 
-    public int GetMoveDifficulty()
+    public float GetMoveDifficulty()
     {
         return moveDifficulty;
     }
